% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{-.tmatrix}
\alias{-.tmatrix}
\title{Unary operator: matrix inverse}
\usage{
\method{-}{tmatrix}(X, Y)
}
\description{
As a unary operator, -X is the same as solve(X), where X is a tmatrix.
In order to negate a tmatrix, use -I(X).
As a binary operator, X-Y has the standard meaning.
}
\examples{
x <- tmatrix(1,2,NULL,3,4)
-x
-I(x)
}
