% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%\%.tmatrix}
\alias{\%\%.tmatrix}
\title{Binary operator: matrix multiplication}
\usage{
\method{\%\%}{tmatrix}(X, Y)
}
\arguments{
\item{X}{tmatrix}

\item{Y}{tmatrix of same shape}
}
\description{
X \%\% Y is the same as Y \%*\% X
}
