% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{+.tmatrix}
\alias{+.tmatrix}
\title{Unary operator: transpose}
\usage{
\method{+}{tmatrix}(X, Y)
}
\arguments{
\item{X}{tmatrix}

\item{Y}{tmatrix of same shape}
}
\description{
As a unary operator, +X is the same as t(X), where X is a tmatrix.
As a binary operator, X+Y has the standard meaning.
}
\examples{
x <- tmatrix(1,2,NULL,3,4,5,6)
+x
x+x
}
