% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tmatrix}
\alias{print.tmatrix}
\title{Print a tmatrix}
\usage{
\method{print}{tmatrix}(x)
}
\arguments{
\item{x}{tmatrix}
}
\description{
Print a tmatrix just like a standard matrix, with a
footer reminding you that it is a tmatrix.
}
