% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmatrix-package.R
\docType{package}
\name{tmatrix-package}
\alias{tmatrix-package}
\title{tmatrix}
\description{
T-matrix method for electromagnetic scattering by particles
}
\details{
R/C++ implementation of the T-matrix method [work in progress] based on other codes.
}
\author{
baptiste Auguie
}
\keyword{package}
\keyword{packagelibrary}

