% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmatrix.R
\name{tmatrix}
\alias{tmatrix}
\title{Create a tmatrix}
\usage{
tmatrix(..., terse = TRUE)
}
\arguments{
\item{...}{elements to include (with NULL terminating the first row)}

\item{terse}{if FALSE, create a matrix (not tmatrix)}
}
\value{
tmatrix (or matrix) with elements and shape as specified in ...
}
\description{
A tmatrix is a matrix with class tmatrix.
Specify elements in row-major order, with the first row terminated by NULL.
}
\examples{
tmatrix(1,2,NULL,
        3,4)
}
